+++ src/main/java/ganymedes01/etfuturum/mixins/thinpanes/MixinRenderBlocks.java
@@ -1,4 +1,4 @@
-package ganymedes01.etfuturum.mixins.glasspane;
+package ganymedes01.etfuturum.mixins.thinpanes;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockPane;
